Client src file Restructure
Moved/Organized components into folders for better organization and file structure
