+++ client/src/components/Views/SpaceX.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import {connect} from 'react-redux';
-import {fetchRockets} from './actions/spacex-action';
-import {fetchCompanyData} from './actions/spacex-action';
-import {fetchLaunchPads} from './actions/spacex-action';
-import {fetchUpcomingLaunches} from './actions/spacex-action';
-import {fetchPastLaunches} from './actions/spacex-action';
+import { connect } from 'react-redux';
+import { fetchRockets } from '../../actions/spacex-action';
+import { fetchCompanyData } from '../../actions/spacex-action';
+import { fetchLaunchPads } from '../../actions/spacex-action';
+import { fetchUpcomingLaunches } from '../../actions/spacex-action';
+import { fetchPastLaunches } from '../../actions/spacex-action';
 
 class SpaceX extends Component {
   componentDidMount() {
